+++ php/Terminus/Models/User.php
@@ -2,7 +2,6 @@
 
 namespace Terminus\Models;
 
-use Terminus\CommandsTerminusCommand;
 use Terminus\Models\Collections\UserOrganizationMemberships;
 use Terminus\Models\TerminusModel;
 use Terminus\Models\Collections\Instruments;
@@ -78,7 +77,7 @@ class User extends TerminusModel {
       $path = 'sites';
     }
     $method   = 'GET';
-    $response = TerminusCommand::request('users', $this->id, $path, $method);
+    $response = $this->request->request('users', $this->id, $path, $method);
     return $response['data'];
   }
 
@@ -90,7 +89,7 @@ class User extends TerminusModel {
   private function setAliases() {
     $path     = 'drush_aliases';
     $method   = 'GET';
-    $response = TerminusCommand::request('users', $this->id, $path, $method);
+    $response = $this->request->request('users', $this->id, $path, $method);
 
     $this->aliases = $response['data']->drush_aliases;
   }