+++ php/Terminus/Request.php
@@ -3,6 +3,7 @@
 namespace Terminus;
 
 use Terminus;
+use Terminus\Utils;
 use GuzzleHttp\Client;
 use GuzzleHttp\Cookie\CookieJar;
 use GuzzleHttp\Psr7\Request as HttpRequest;
@@ -20,6 +21,202 @@ use Terminus\Exceptions\TerminusException;
 class Request {
 
   /**
+   * A list of fields not to display values for in output
+   * TODO: Move this logic to the logger
+   */
+  protected static $blacklist = array('password');
+
+  /**
+   * Download file from target URL
+   *
+   * @param [string] $url    URL to download from
+   * @param [string] $target Target file's name
+   * @return [boolean] True if download succeeded
+   */
+  public static function download($url, $target) {
+    if (file_exists($target)) {
+      throw new TerminusException(
+        'Target file {target} already exists.',
+        compact('target')
+      );
+    }
+
+    try {
+      $client   = new Client();
+      $response = $client->request('GET', $url, array('sink' => $target));
+    } catch (\Exception $e) {
+      throw new TerminusException($e->getMessage(), array(), 1);
+    }
+    return true;
+  }
+
+  /**
+   * Make a request to the Dashbord's internal API
+   *
+   * @param [string] $path    API path (URL)
+   * @param [array]  $options Options for the request
+   *   [string] method GET is default
+   *   [mixed]  data   Native PHP data structure (e.g. int, string array, or
+   *     simple object) to be sent along with the request. Will be encoded as
+   *     JSON for you.
+   * @return [array] $return
+   */
+  public function pagedRequest($path, $options = array()) {
+    $limit = 100;
+    if (isset($options['limit'])) {
+      $limit = $options['limit'];
+    }
+
+    //$results is an associative array so we don't refetch
+    $results  = array();
+    $finished = false;
+    $start    = null;
+
+    while (!$finished) {
+      $paged_path = $path . '?limit=' . $limit;
+      if ($start) {
+        $paged_path .= '&start=' . $start;
+      }
+
+      $resp = $this->simpleRequest($paged_path);
+
+      $data = $resp['data'];
+      if (count($data) > 0) {
+        $start = end($data)->id;
+
+        //If the last item of the results has previously been received,
+        //that means there are no more pages to fetch
+        if (isset($results[$start])) {
+          $finished = true;
+          continue;
+        }
+
+        foreach ($data as $item) {
+          $results[$item->id] = $item;
+        }
+      } else {
+        $finished = true;
+      }
+    }
+
+    $return = array('data' => array_values($results));
+    return $return;
+  }
+
+  /**
+   * Make a request to the Pantheon API
+   *
+   * @param [string] $realm   Permissions realm for data request (e.g. user,
+   *   site organization, etc. Can also be "public" to simply pull read-only
+   *   data that is not privileged.
+   * @param [string] $uuid    The UUID of the item in the realm to access
+   * @param [string] $path    API path (URL)
+   * @param [string] $method  HTTP method to use
+   * @param [mixed]  $options A native PHP data structure (e.g. int, string,
+   *   array, or stdClass) to be sent along with the request
+   * @return [array] $data
+   */
+  public function request(
+    $realm,
+    $uuid,
+    $path    = false,
+    $method  = 'GET',
+    $options = array()
+  ) {
+    $logger = Terminus::getLogger();
+
+    try {
+      $url = Endpoint::get(
+        array(
+          'realm' => $realm,
+          'uuid'  => $uuid,
+          'path'  => $path,
+        )
+      );
+      $logger->debug('Request URL: ' . $url);
+      $response = $this->send($url, $method, $options);
+
+      $data = array(
+        'data'        => json_decode($response->getBody()->getContents()),
+        'headers'     => $response->getHeaders(),
+        'status_code' => $response->getStatusCode(),
+      );
+      return $data;
+    } catch (GuzzleHttp\Exception\BadResponseException $e) {
+      $response = $e->getResponse();
+      throw new TerminusException($response->getBody(true));
+    } catch (GuzzleHttp\Exception\HttpException $e) {
+      $request = $e->getRequest();
+      $sanitized_request = Utils\stripSensitiveData(
+        (string)$request,
+        $this->blacklist
+      );
+      throw new TerminusException(
+        'API Request Error. {msg} - Request: {req}',
+        array('req' => $sanitized_request, 'msg' => $e->getMessage())
+      );
+    } catch (Exception $e) {
+      throw new TerminusException(
+        'API Request Error: {msg}',
+        array('msg' => $e->getMessage())
+      );
+    }
+  }
+
+  /**
+   * Simplified request method for Pantheon API
+   *
+   * @param [string] $path    API path (URL)
+   * @param [array]  $options Options for the request
+   *   [string] method GET is default
+   *   [mixed]  data   Native PHP data structure (e.g. int, string array, or
+   *     simple object) to be sent along with the request. Will be encoded as
+   *     JSON for you.
+   * @return [array] $data
+   */
+  public function simpleRequest($path, $options = array()) {
+    $method = 'get';
+    if (isset($options['method'])) {
+      $method = $options['method'];
+      unset($options['method']);
+    }
+
+    $url = $path;
+    if (!isset($options['absolute_url']) || !$options['absolute_url']) {
+      $url = sprintf(
+        '%s://%s:%s/api/%s',
+        TERMINUS_PROTOCOL,
+        TERMINUS_HOST,
+        TERMINUS_PORT,
+        $path
+      );
+    }
+
+    if (Session::getValue('session')) {
+      $options['cookies'] = array(
+        'X-Pantheon-Session' => Session::getValue('session')
+      );
+    }
+
+    try {
+      Terminus::getLogger()->debug('URL: {url}', compact('url'));
+      $response = $this->send($url, $method, $options);
+    } catch (GuzzleHttp\Exception\BadResponseException $e) {
+      throw new TerminusException(
+        'API Request Error: {msg}',
+        array('msg' => $e->getMessage())
+      );
+    }
+
+    $data = array(
+      'data'        => json_decode($response->getBody()->getContents()),
+      'headers'     => $response->getHeaders(),
+      'status_code' => $response->getStatusCode(),
+    );
+    return $data;
+  }
+
+  /**
    * Sends a request to the API
    *
    * @param [string] $uri        URL for API request
@@ -27,10 +224,10 @@ class Request {
    * @param [array]  $arg_params Request parameters
    * @return [GuzzleHttp\Message\Response] $response
    */
-  public static function send($uri, $method, array $arg_params = array()) {
+  private function send($uri, $method, array $arg_params = array()) {
     $extra_params = array(
       'headers'         => array(
-        'User-Agent'    => self::userAgent(),
+        'User-Agent'    => $this->userAgent(),
         'Content-type'  => 'application/json',
       ),
     );
@@ -46,8 +243,8 @@ class Request {
 
     $client = new Client(
       array(
-        'base_uri' => self::getBaseUri(),
-        'cookies'  => self::fillCookieJar($params)
+        'base_uri' => $this->getBaseUri(),
+        'cookies'  => $this->fillCookieJar($params)
       )
     );
     unset($params['cookies']);
@@ -71,36 +268,12 @@ class Request {
   }
 
   /**
-   * Download file from target URL
-   *
-   * @param [string] $url    URL to download from
-   * @param [string] $target Target file's name
-   * @return [boolean] True if download succeeded
-   */
-  static function download($url, $target) {
-    if (file_exists($target)) {
-      throw new TerminusException(
-        'Target file {target} already exists.',
-        compact('target')
-      );
-    }
-
-    try {
-      $client   = new Client();
-      $response = $client->request('GET', $url, array('sink' => $target));
-    } catch (\Exception $e) {
-      throw new TerminusException($e->getMessage(), array(), 1);
-    }
-    return true;
-  }
-
-  /**
    * Sets up and fills a cookie jar
    *
    * @param [array] $params Request data to fill jar with
    * @return [GuzzleHttp\Cookie\CookieJar] $jar
    */
-  static function fillCookieJar($params) {
+  private function fillCookieJar($params) {
     $jar     = new CookieJar();
     $cookies = array();
     if ($session = Session::instance()->get('session', false)) {
@@ -118,7 +291,7 @@ class Request {
    *
    * @return [string] $base_uri
    */
-  static function getBaseUri() {
+  private function getBaseUri() {
     $base_uri = sprintf(
       '%s://%s:%s',
       TERMINUS_PROTOCOL,
@@ -133,7 +306,7 @@ class Request {
    *
    * @return [string] $agent
    */
-  static function userAgent() {
+  private function userAgent() {
     $agent = sprintf(
       'Terminus/%s (php_version=%s&script=%s)',
       constant('TERMINUS_VERSION'),