+++ CHANGELOG.md
@@ -10,6 +10,9 @@ All notable changes to this project starting with the 0.6.0 release will be docu
 - `site backups get` no longer errs when there are no backups. (#690)
 - Interactive commands' environment menus now consistently include multidev environments, where applicable. (#701)
 
+### Changed
+- Extricated the request logic from TerminusCommand class and moved it to the Request class. (#704)
+
 ##[0.9.3] - 2015-11-17
 ### Added
 - `site environments` now includes data on whether environment is initialized yet. (#654)